Add Shortcode functionality
This introduces a base shortcode class `VGSR_Shortcode` which handles
basic registration, registration for Shortcode UI, and parsing of
attributes.

The first implementation of this logic is the [vgsr-subpages] shortocde
through the `VGSR_Shortcode_Subpages` class.

Although the plugin name doesn't provide a place for shortcodes, it seemed
the best place to include them, for now.

See issue 16 for vgsr/zeta.